backend/requirements.txt
fastapi
uvicorn
motor
google-generativeai
python-dotenv
bcrypt
pyjwt
python-multipart
pydantic